Add files via upload
This is a dynamic Data Userform Project which includes a multipage. This multipage is having two tabs one is personal and the another one is Job related.In personal Tab there are 4 buttons of submit,reset,display and exit. The submit button will submit all the records based on the user input. Reset button will reset the data in the textboxes and display button will display the record based on the user selection from the combo box